+++ src/Lens/Traits/HttpConnectionLens.php
@@ -2,17 +2,29 @@
 namespace HiFolks\LaraLens\Lens\Traits;
 
 
+use HiFolks\LaraLens\Console\LaraLensCommand;
 use HiFolks\LaraLens\ResultLens;
 use Illuminate\Support\Facades\Http;
 
 trait HttpConnectionLens
 {
-    public function getConnections()
+    public function getConnections($checkPath)
     {
         $results = new ResultLens();
-        $url = config("app.url");
+        $app_url = config("app.url");
         $results->add(
-            "Connect to URL",
+            "app.url configuration",
+            $app_url
+        );
+        $url = url($checkPath);
+
+
+        $results->add(
+            "url()->full()",
+            url()->full()
+        );
+        $results->add(
+            "Connection HTTP URL",
             $url
         );
 
@@ -23,11 +35,17 @@ trait HttpConnectionLens
                 $response->status()
             );
             if ($response->failed()) {
+                $checkUrlHint = "Check APP_URL '" . $app_url . "' in .env file ";
+
+                if ($checkPath !== "") {
+                    $checkUrlHint .= " or check this path: '" . $checkPath . "' in routing file (routes/web.php)";
+                }
                 $this->checksBag->addWarningAndHint(
                     "Connection HTTP Status",
-                    "Connection response not 20x, status code: " . $response->status(),
-                    "Check this URL: " . $url . " in .env file APP_URL"
+                    "Connection response not 20x, status code: " . $response->status() . " for ".$url,
+                    $checkUrlHint
                 );
+
 
             }
 
@@ -40,7 +58,7 @@ trait HttpConnectionLens
             $this->checksBag->addErrorAndHint(
                 "Connection HTTP Status",
                 "Connection Error: " . $e->getMessage(),
-                "Check this URL: " . $url . " in .env file APP_URL"
+                "Check this URL: " . $app_url . " in .env file APP_URL"
             );
         }
         return $results;